Basic refactor to make the code more robust
